Fix non-spin modules for coupling
