+++ sympy/physics/quantum/qapply.py
@@ -137,7 +137,9 @@ def qapply_Mul(e, **options):
             return qapply(e._new_rawargs(*args)*comm*rhs, **options)
 
     # Apply tensor products of operators to states
-    if isinstance(lhs, TensorProduct) and any([isinstance(arg,Operator) for arg in lhs.args]) and isinstance(rhs, TensorProduct):
+    if isinstance(lhs, TensorProduct) and all([isinstance(arg,Operator) or arg == 1 for arg in lhs.args]) and \
+            isinstance(rhs, TensorProduct) and all([isinstance(arg,State) or arg == 1 for arg in rhs.args]) and \
+            len(lhs.args) == len(rhs.args):
         result = TensorProduct(*[qapply(lhs.args[n]*rhs.args[n], **options) for n in range(len(lhs.args))]).expand(tensorproduct=True)
         return qapply_Mul(e._new_rawargs(*args), **options)*result
